requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
dj-rest-auth>=6.0,<7.0
django-allauth>=65.0,<66.0
djangorestframework-simplejwt>=5.3,<6.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
gunicorn>=23.0,<24.0
whitenoise>=6.7,<7.0
django-filter>=24.3,<25.0
dj-database-url>=2.2,<3.0
